write zip code
