+++ packages/server/src/index.ts
@@ -22,6 +22,8 @@ export * as TodosRepository from "./TodosRepository.js";
 
 export * as Users from "./Users.js";
 
+export * as Zip from "./Zip.js";
+
 export * as ACarFullBackup from "./importers/ACarFullBackup.js";
 
 export * as main from "./main.js";
